requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
pymongo==3.11.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
